Fix flaky "race successful with wait" test
This used to fail sometimes with the second source not getting rid of the listener.
Occurs when the second listener call won the race, which is quite rare.

Due to the contract of the Source, it will remove the listener, but the second
listener is called bare, so cleanup does not happen.