Avoid `native` name clash
In this file we declare the package `gears.async.native` while also
importing `scala.scalanative.runtime.Continuations` as `native`.

In Scala up until 3.3.1, `native.suspend` and such looks up on the renamed
import and the code compiles. Scala 3.4.0 looks for the name in the local
package, and hence does not compile.