Experiment: Only allow `Async.Spawnable` to spawn runnable futures
The goal is to disallow spawning dangling Futures from `using Async` functions.
`Async.Spawnable` is an opaque alias of `Async`, defined as a subtype of `Async`,
obtained by explicitly "upgrading" it through `Async.spawning` (which works similarly
to `Async.group`, so futures are all cancelled after) -- or automatically given
through `Async.blocking` or `Future.apply`.

The `Async.Spawnable`-taking functions (signalling usage of dangling futures)
should follow the hacky signature of `Future.apply`:

> def apply[T](body: Async.Spawnable ?=> T)
>   (using async: Async, spawn: Async.Spawnable & async.type): T

to ensure that the given `Async` instance (which is usually synthesized
to be the innermost context) is the same instance as the `Async.Spawnable`
instance. It happens quite often (especially when nesting `Async` contexts)
that these don't match:

> extension[T] (s: Seq[T])
>   def parallelMap[U](f: T => Async ?=> U)(using Async): Seq[U]
>
> Async.blocking: // Async.Spawnable here...
>   val seq = Seq(1, 2, 3, 4, 5)
>     .parallelMap: n => // Async here...
>        Async.select(
>           Future(doSomethingAsync(n)) handle Some(_),     // oops, spawned by Async.blocking
>           Future(Async.sleep(1.minute)) handle _ => None, // oops, spawned by Async.blocking
>        )
>   // oops, leaking all the futures...

with the `Future.apply` signature as above, this does not happen and will
give a compile time error.